convert numerical value to float.
An attempt to convert the string numerical values to a float for the sum_numbers function.

+++ python_edu/ex_7_2/ex_7_2.py
@@ -17,6 +17,6 @@ def sum_numbers(numbers):
         return 0
     return numbers[0] + sum_numbers(numbers[1:])
 
-sum_numbers(fl)
+sum_numbers(fl.float())
 
 print("Done")
